Homework 2 added

+++ my-app/backend/db/migrate/20230622001605_create_users.rb
@@ -5,7 +5,7 @@ class CreateUsers < ActiveRecord::Migration[7.0]
       t.string :pin
       t.boolean :is_clocked_in
       t.boolean :is_admin
-      t.integer :hours_worked
+      t.float :hours_worked
       t.datetime :time_in
 
       t.timestamps
